requirements.txt
# Application
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
openai-whisper==20240930
pydantic==2.10.3
python-multipart==0.0.19
numpy<2

# Testing
pytest==8.3.4
httpx==0.28.1
